Create test utils for e2e tests
